Move add_non_indexed to reader
