+++ src/napari_defdap/_reader.py
@@ -12,6 +12,21 @@ import yaml
 from defdap import hrdic, ebsd
 
 
+def add_non_indexed(seg, time_axis=0):
+    ndim = seg.ndim
+    non_indexed = seg <= 0
+    axes = tuple(i for i in range(ndim) if i != time_axis)
+    max_label = np.max(seg, axis=axes)
+    labeled = np.zeros_like(seg)
+    for i in range(seg.shape[time_axis]):
+        idx_ = [slice(None),] * ndim
+        idx_[time_axis] = i
+        idx = tuple(idx_)
+        labeled[idx] = ndi.label(non_indexed[idx])[0] + max_label[i]
+    output = np.where(non_indexed, labeled, seg)
+    return output
+
+
 def _look_for_ebsd(path):
     directory = os.path.split(path)[0]
     files = [os.path.join(directory, os.path.splitext(fn)[0])
